+++ src/v1/ErrorTable/PackageErrorTable.ts
@@ -31,15 +31,16 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { httpStatusCodeFrom } from "@ganbarodigital/ts-lib-http-types/lib/v1";
-import { packageNameFrom } from "@ganbarodigital/ts-lib-packagename/lib/v1";
-
+import { HttpStatusCode } from "../copied/HttpStatusCode";
+import { PackageName } from "../copied/PackageName";
 import { ErrorTableTemplateWithNoExtraData } from "../ErrorTableTemplate";
 import { ExtraDataTemplate, NoExtraDataTemplate } from "../ExtraData";
 import { ErrorTable } from "./ErrorTable";
 import { UnreachableCodeTemplate } from "./UnreachableCode";
 
-export const PACKAGE_NAME = packageNameFrom("@ganbarodigital/ts-lib-error-reporting/lib/v1");
+// we can't use `packageNameFrom()` from `ts-lib-packagename` here, because
+// it creates a circular dependency that stops unit tests from compiling
+export const PACKAGE_NAME = "@ganbarodigital/ts-lib-error-reporting/lib/v1" as PackageName;
 
 /**
  * the ErrorTable for the package `@ganbarodigital/ts-lib-error-reporting`
@@ -58,7 +59,7 @@ export class PackageErrorTable implements ErrorTable {
     public "unreachable-code": UnreachableCodeTemplate = {
         packageName: PACKAGE_NAME,
         errorName: "unreachable-code",
-        status: httpStatusCodeFrom(500),
+        status: 500 as HttpStatusCode,
         detail: "this code should never execute",
         extra: {
             logsOnly: {